refactor(chart): rename date utility to format
